+++ test/testSchemaGenerator.js
@@ -619,6 +619,35 @@ describe('generating schema from shorthand', () => {
     assert.doesNotThrow(makeExecutableSchema.bind(null, { typeDefs: short, resolvers: rf, resolverValidationOptions: { requireResolversForNonScalar: false } }), SchemaError); // eslint-disable-line max-len
   });
 
+  it('throws for any missing field if `resolverValidationOptions.requireResolversForAllFields` = true', () => { // eslint-disable-line max-len
+    const typeDefs = `
+    type Bird {
+      id: ID
+    }
+    type Query {
+      bird: Bird
+    }
+    schema {
+      query: Query
+    }`;
+
+    function assertFieldError(errorMatcher, resolvers) {
+      assert.throws(() => makeExecutableSchema({ typeDefs, resolvers, resolverValidationOptions: { requireResolversForAllFields: true } }), SchemaError, errorMatcher); // eslint-disable-line max-len
+    }
+
+    assertFieldError(null, {});
+    assertFieldError('Query.bird', {
+      Bird: {
+        id: bird => bird.id,
+      },
+    });
+    assertFieldError('Bird.id', {
+      Query: {
+        bird: () => ({ id: '123' }),
+      },
+    });
+  });
+
   it('throws an error if a resolve field cannot be used', (done) => {
     const shorthand = `
       type BirdSpecies {