+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### vNEXT
 
+* Add `requireResolversForAllFields` resolver validation option ([@nevir](https://github.com/nevir) in PR)
+
 ### v0.6.4
 * Make mocking partial objects match expected behavior ([@sebastienbarre](https://github.com/sebastienbarre) in [#96](https://github.com/apollostack/graphql-tools/pull/96))
 * Improved behavior when mocking interfaces & unions ([@itajaja](https://github.com/itajaja) in [#102](https://github.com/apollostack/graphql-tools/pull/102))